feat: restructure dashboard layout and add tasks page with task management features
